Implement user-sport enrollement and disenrollement functionality

+++ src/modules/user/user.service.ts
@@ -1,13 +1,16 @@
-import { HttpStatus, Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CreateUserDto } from './dtos/createUser.dto';
 import { ReadUserDto } from './dtos/readUser.dto';
 import { User } from './user.entity';
 import { Repository } from 'typeorm';
-import { AgeGroup } from '../age-group/ageGroup';
+import { AgeGroup } from '../ageGroup/ageGroup';
+import { Sport } from '../sport/sport.enitity';
+import { idException } from 'src/exceptions/idException';
 
 @Injectable()
 export class UserService {
+  private readonly maxSportsQuantity: number = 2;
   constructor(
     @InjectRepository(User) private readonly userRepository: Repository<User>,
   ) {}
@@ -35,20 +38,67 @@ export class UserService {
     );
     return await this.userRepository.save(user);
   }
-  async updateUser(id: number, item: CreateUserDto): Promise<CreateUserDto> {
+  async enrollUserInSport(
+    userId: number,
+    sport: Sport,
+  ): Promise<HttpStatus.ACCEPTED | HttpException> {
+    const user = await this.userRepository.findOne({
+      where: {
+        id: userId,
+      },
+      relations: {
+        sports: true,
+      },
+    });
+    if (user) {
+      if (user.sports.length < this.maxSportsQuantity) {
+        user.sports.push(sport);
+        await this.userRepository.save(user);
+        return HttpStatus.ACCEPTED;
+      } else
+        return new HttpException(
+          `User can enroll in a maximum of ${this.maxSportsQuantity} sports`,
+          HttpStatus.BAD_REQUEST,
+        );
+    }
+    return idException;
+  }
+
+  async disenrollUserFromSport(
+    userId: number,
+    sportId: number,
+  ): Promise<HttpStatus.ACCEPTED> {
+    const user = await this.userRepository.findOneOrFail({
+      where: {
+        id: userId,
+      },
+      relations: {
+        sports: true,
+      },
+    });
+    user.sports = user.sports.filter(
+      (userSport) => userSport.id !== Number(sportId),
+    );
+    await this.userRepository.save(user);
+    return HttpStatus.ACCEPTED;
+  }
+  async updateUser(id: number, item: CreateUserDto): Promise<User> {
     const user = await this.userRepository.findOneByOrFail({ id: id });
     const updatedUser = await this.userRepository.save({
       id: user.id,
       name: item.name,
       age: item.age,
       email: item.email,
-      sports: item.sports,
+      sports: user.sports,
     });
     return updatedUser;
   }
   async deleteUser(id: number): Promise<HttpStatus.ACCEPTED> {
-    await this.userRepository.delete(id);
-    return HttpStatus.ACCEPTED;
+    const user = await this.userRepository.findOneByOrFail({ id: id });
+    if (user) {
+      await this.userRepository.delete(id);
+      return HttpStatus.ACCEPTED;
+    }
   }
   calculateAgeGroup(age: number): AgeGroup {
     switch (true) {
